requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
Django==4.2.2
django-cors-headers==4.1.0
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
jwt==1.3.1
pycparser==2.21
pygraphviz==1.11
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.7.0
urllib3==2.0.3
